+++ _unittests/ut_helpgen/test_module_c.py
@@ -149,6 +149,8 @@ class TestModuleC(ExtTestCase):
         indexes = {}
         add_file_rst(temp, store_obj, actions, fLOG=fLOG,
                      rootrep=("module_c.", ""), indexes=indexes)
+        self.assertNotEmpty(store_obj)
+        self.assertEqual(len(store_obj), 4)
         if len(actions) != 2:
             raise Exception("{0}\n{1}".format(
                 len(actions), "\n".join(str(_) for _ in actions)))
@@ -207,6 +209,8 @@ class TestModuleC(ExtTestCase):
         indexes = {}
         add_file_rst(temp2, store_obj, actions, fLOG=fLOG,
                      rootrep=("stdchelper_demo.", ""), indexes=indexes)
+        self.assertNotEmpty(store_obj)
+        self.assertEqual(len(store_obj), 1)
         if len(actions) not in (3, 4):
             raise Exception("{0}\n{1}".format(
                 len(actions), "\n".join(str(_) for _ in actions)))